deleted things, that are incompatible with ir incoders setup, everythingworks perfect

+++ ROSArduinoBridge/motor_driver.h
@@ -15,5 +15,5 @@ extern volatile bool right_reverse;
 #endif
 
 void initMotorController();
-void setMotorSpeed(int i, int spd);
-void setMotorSpeeds(int leftSpeed, int rightSpeed);
+void setMotorSpeed(float i, float spd);
+void setMotorSpeeds(float leftSpeed, float rightSpeed);
